First version of doc blocks migration guide
