+++ code/ui/blocks/src/blocks/Story.stories.tsx
@@ -29,22 +29,6 @@ export const OfWithMeta: Story = {
   },
 };
 
-const blocksAwareId = `${
-  import.meta.env.STORYBOOK_BLOCKS_ONLY === 'true' ? '' : 'storybook-blocks-'
-}example-button--primary`;
-
-export const Id: Story = {
-  args: {
-    id: blocksAwareId,
-  },
-};
-
-export const Name: Story = {
-  args: {
-    name: 'Secondary',
-  },
-};
-
 export const SimpleSizeTest: Story = {
   render: () => {
     return (
